Templated Forms

+++ ContactManager/Migrations/ContactContextModelSnapshot.cs
@@ -29,7 +29,7 @@ namespace ContactManager.Migrations
 
                     b.Property<string>("Email");
 
-                    b.Property<int?>("NameID");
+                    b.Property<long?>("NameID");
 
                     b.HasKey("ID");
 
@@ -40,7 +40,7 @@ namespace ContactManager.Migrations
 
             modelBuilder.Entity("ContactManager.Model.Name", b =>
                 {
-                    b.Property<int>("ID")
+                    b.Property<long>("ID")
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("First");
@@ -57,7 +57,7 @@ namespace ContactManager.Migrations
                     b.Property<long>("ID")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<int?>("NameID");
+                    b.Property<long?>("NameID");
 
                     b.Property<string>("Telephone");
